% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lee.R
\name{lee}
\alias{lee}
\title{lee.R
Lee. et al's method of approximating the mixture distribution of Q to match the theoretical distribution
this is based on Liu. et al's method
Helper function for getting the parameters for the test statistics approximation}
\usage{
lee(c)
}
\arguments{
\item{c}{A vector of 4 elements c=c(c1,c2,c3,c4).}
}
\value{
A list with the elements:
\item{l}{The degree of freedom of chi-square distribution.}
\item{delta}{The non-centroality parameter of chi-square distribution.}
\item{mux}{The mean of chi-square distribution}
\item{sigmax}{The standard deviation of chi-square distribution}
\item{muq}{The mean of test statistics.}
\item{muq}{The standard deviation of test statistics.}
}
\description{
lee.R
Lee. et al's method of approximating the mixture distribution of Q to match the theoretical distribution
this is based on Liu. et al's method
Helper function for getting the parameters for the test statistics approximation
}
\examples{
c<-c(1,2,3,4)
lee(c)
# Should return l=2.220446e-15, delta=1, mux=1, sigmax=2, muq=1, sigmaq=2.
}
