% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powerC_derive.R
\name{powerC_derive}
\alias{powerC_derive}
\title{powerC_derive.R
calculate theoretical power for SKAT test for continuous outcome, uncorrelated genotypes.}
\usage{
powerC_derive(k, n, alpha, p, list1, list2)
}
\arguments{
\item{k}{Number of SNPs.}

\item{n}{Number of subjects.}

\item{alpha}{Significance level.}

\item{p}{Minor allele frequency.}

\item{list1}{Vector of effect sizes for the first alternative allele in the multi-allelic position.}

\item{list2}{Vector of effect sizes for the second alternative allele in the multi-allelic position.}
}
\value{
\code{power}:Vector of derived power for SKAT test.
}
\description{
powerC_derive.R
calculate theoretical power for SKAT test for continuous outcome, uncorrelated genotypes.
}
\examples{
beta1<-c(-0.9, -0.75, -0.5, -0.25, -0.1)
beta2<-c(rep(0.5, 5))
powerC_derive(k=10,n=2000,alpha=0.05,p=0.01,list1=beta1,list2=beta2)
# Should return 0.9991246 0.9912256 0.8957736 0.6698220 0.5704163
}
