% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powerC_deriveCor.R
\name{powerC_deriveCor}
\alias{powerC_deriveCor}
\title{powerC_deriveCor.R
calculate theoretical power for SKAT test for continuous outcome, correlated genotypes}
\usage{
powerC_deriveCor(k, n, alpha, p, rho, list1, list2)
}
\arguments{
\item{k}{Number of SNPs.}

\item{n}{Number of subjects.}

\item{alpha}{Significance level.}

\item{p}{Minor allele frequency.}

\item{rho}{Correlation coefficient between SNPs.}

\item{list1}{Vector of effect sizes for the first alternative allele in the multi-allelic position.}

\item{list2}{Vector of effect sizes for the second alternative allele in the multi-allelic position.}
}
\value{
\code{power}:Vector of derived power for SKAT test.
}
\description{
powerC_deriveCor.R
calculate theoretical power for SKAT test for continuous outcome, correlated genotypes
}
\examples{
beta1<-c(-0.9, -0.75, -0.5, -0.25, -0.1)
beta2<-c(rep(0.5, 5))
powerC_deriveCor(k=10,n=2000,alpha=0.05,p=0.01,rho=0.1,list1=beta1,list2=beta2)
# Should return 0.9962518 0.9717550 0.7914034 0.5562763 0.5183587
}
