% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powerC_meanCor.R
\name{powerC_meanCor}
\alias{powerC_meanCor}
\title{powerC_meanCor.R
For continuous outcome and correalted genotypes, calculate simulated power for combined and seperated modles using SKAT, and also calculate the theoretical power using  Lee's method}
\usage{
powerC_meanCor(n, k, p, alpha, rho, runs, list1, list2)
}
\arguments{
\item{n}{Number of subjects.}

\item{k}{Number of SNPs.}

\item{p}{Minor allele frequency.}

\item{alpha}{Significance level.}

\item{rho}{Correlatoin coefficient between SNPs.}

\item{runs}{Number of simulations.}

\item{list1}{Vector of effect sizes for the first alternative allele in the multi-allelic position.}

\item{list2}{Vector of effect sizes for the second alternative allele in the multi-allelic position.}
}
\value{
\code{powerE,powerSeparated,powerCombined}:Vector of three powers for SKAT test.
}
\description{
powerC_meanCor.R
For continuous outcome and correalted genotypes, calculate simulated power for combined and seperated modles using SKAT, and also calculate the theoretical power using  Lee's method
}
\examples{
beta1=-0.5
beta2=0.75
powerC_meanCor(k=10,n=200,alpha=0.05,p=0.1,rho=0.1,runs=10,list1=beta1,list2=beta2)
# Should return powerE=0.9650208,powerSeparated=0.7,powerCombined=0.1
}
