% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powerD_derive.R
\name{powerD_derive}
\alias{powerD_derive}
\title{powerD_derive.R
calculate the  theoretical power for SKAT test for binary outcome, uncorrelated genotypes}
\usage{
powerD_derive(kk, n, p, alpha, list1, list2)
}
\arguments{
\item{kk}{Number of SNPs.}

\item{n}{Number of subjects.}

\item{p}{Minor allele frequency.}

\item{alpha}{Significance level.}

\item{list1}{Vector of effect sizes for the first alternative allele in  multi-allelic position.}

\item{list2}{Vector of effect sizes for the second alternative allele in  multi-allelic position.}
}
\value{
\code{power}:Vector of derived power for SKAT test.
}
\description{
powerD_derive.R
calculate the  theoretical power for SKAT test for binary outcome, uncorrelated genotypes
}
\examples{
beta1<-c(-0.9, -0.75, -0.5, -0.25, -0.1)
beta2<-c(rep(0.5, 5))
powerD_derive(k=10,n=2000,alpha=0.05,p=0.01,list1=beta1,list2=beta2)
# Should return 0.9991246 0.9912256 0.8957736 0.6698220 0.5704163
}
