% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powerD_deriveCor.R
\name{powerD_deriveCor}
\alias{powerD_deriveCor}
\title{powerD_deriveCor.R
calculate theoretical power for SKAT test for binary outcome, correlated genotypes}
\usage{
powerD_deriveCor(kk, n, alpha, p, rho, list1, list2)
}
\arguments{
\item{kk}{Number of SNPs.}

\item{n}{Number of subjects.}

\item{alpha}{Significance level.}

\item{p}{Minor allele frequency.}

\item{rho}{Correlation coefficient between SNPs.}

\item{list1}{Vector of effect sizes for the first alternative allele in  multi-allelic position.}

\item{list2}{Vector of effect sizes for the second alternative allele in  multi-allelic position.}
}
\value{
\code{power}:Vector of derived power for SKAT test.
}
\description{
powerD_deriveCor.R
calculate theoretical power for SKAT test for binary outcome, correlated genotypes
}
\examples{
beta1<-c(-0.9, -0.75, -0.5, -0.25, -0.1)
beta2<-c(rep(0.5, 5))
powerD_deriveCor(k=10,n=2000,alpha=0.05,p=0.01,rho=0.15,list1=beta1,list2=beta2)
# Should return 0.25564418 0.15141926 0.08757907 0.13321503 0.21503152
}
