% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powerD_mean.R
\name{powerD_mean}
\alias{powerD_mean}
\title{powerD_meanCor.R
For binary outcome and uncorrealted genotypes, calculate simulated power for combined and seperated modles using SKAT, and also calculate the theoretical power using  Lee's method}
\usage{
powerD_mean(k, n, alpha, p, runs, list1, list2)
}
\arguments{
\item{k}{Number of SNPs.}

\item{n}{Number of subjects.}

\item{alpha}{Significance level.}

\item{p}{Minor allele frequency.}

\item{runs}{Number of simulations.}

\item{list1}{Vector of effect sizes for the first alternative allele in the multi-allelic position.}

\item{list2}{Vector of effect sizes for the second alternative allele in the multi-allelic position.}
}
\value{
\code{powerE,powerSeparated,powerCombined}:Vector of three powers for SKAT test.
}
\description{
powerD_meanCor.R
For binary outcome and uncorrealted genotypes, calculate simulated power for combined and seperated modles using SKAT, and also calculate the theoretical power using  Lee's method
}
\examples{
beta1=-0.5
beta2=0.75
powerD_mean(k=10,n=2000,alpha=0.05,p=0.1,runs=10,list1=beta1,list2=beta2)
# Should return powerE= 0.9999946,powerSeparated=1,powerCombined=0.3
}
