% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powerD_meanCor.R
\name{powerD_meanCor}
\alias{powerD_meanCor}
\title{powerD_meanCor.R
For binary outcome and correlated genotypes, calculate simulated power for combined and seperated modles using SKAT, and also calculate the theoretical power using  Lee's method}
\usage{
powerD_meanCor(n, k, alpha, p, rho, runs, list1, list2)
}
\arguments{
\item{n}{Number of subjects.}

\item{k}{Number of SNPs.}

\item{alpha}{Significance level.}

\item{p}{Minor allele frequency.}

\item{rho}{Correlation coefficient between SNPs.}

\item{runs}{Number of simulations.}

\item{list1}{Vector of effect sizes for the first alternative allele in the multi-allelic position.}

\item{list2}{Vector of effect sizes for the second alternative allele in the multi-allelic position.}
}
\value{
\code{powerE,powerSeparated,powerCombined}:Vector of three powers for SKAT test.
}
\description{
powerD_meanCor.R
For binary outcome and correlated genotypes, calculate simulated power for combined and seperated modles using SKAT, and also calculate the theoretical power using  Lee's method
}
\examples{
beta1=-0.5
beta2=0.75
powerD_meanCor(k=10,n=2000,alpha=0.05,p=0.1,rho=0.1,runs=10,list1=beta1,list2=beta2)
# Should return powerE= 0.2859357,powerSeparated=0.2,powerCombined=0
}
